Swap to retrieve wishlist from API
It doesn't include as much info, but we don't really care about most
of it. The only thing we really want is the game title, so we've also
overhauled how that data is retrieved, so we only need to do so once
per run. Otherwise we would have had a bunch of duplicated code.