Changed App Info retrieval
Steam seems to have removed the old endpoint for getting all the game
names, and now it only has one for games actually listed on the store.
With that in mind, we might as well just query the store API for app
info instead, which involves a lot less irrelevant data, but a lot more
queries.

Also added a few other bits and pieces of info to profile output.